+++ changelog.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+## Version 5.2.10 (April 29th 2015)
+* Add custom header getter
 * Use `application/javascript` for .js attachments
 * Improve RFC2821 compliance for timelimits, especially for end-of-data
 * Add Azerbaijani translations (Thanks to @mirjalal)
@@ -27,6 +29,9 @@
 * Enable TLS encryption automatically if the server offers it
 * Provide detailed errors when individual recipients fail
 * Report more errors when connecting
+* Add extras classes to composer classmap
+* Expose stream_context_create options via new SMTPOptions property
+* Automatic encoding switch to quoted-printable if message lines are too long
 
 ## Version 5.2.9 (Sept 25th 2014)
 * **Important: The autoloader is no longer autoloaded by the PHPMailer class**